+++ rustfst/src/fst_impls/const_fst/converters.rs
@@ -1,0 +1,30 @@
+use crate::fst_impls::const_fst::data_structure::ConstState;
+use crate::fst_impls::{ConstFst, VectorFst};
+use crate::fst_traits::ExpandedFst;
+use crate::semirings::Semiring;
+
+impl<W: Semiring + 'static> From<VectorFst<W>> for ConstFst<W> {
+    fn from(ifst: VectorFst<W>) -> Self {
+        let mut const_states = Vec::with_capacity(ifst.num_states());
+        // TODO: Add a num_arcs in VectorFst to avoid iterating through all the states..
+        let mut const_arcs = Vec::with_capacity(ifst.states.iter().map(|s| s.arcs.len()).sum());
+        let mut pos = 0;
+        for s in ifst.states {
+            const_states.push(ConstState {
+                final_weight: s.final_weight,
+                pos,
+                narcs: s.arcs.len(),
+            });
+
+            pos += s.arcs.len();
+
+            const_arcs.extend(s.arcs.into_iter());
+        }
+
+        ConstFst {
+            states: const_states,
+            arcs: const_arcs,
+            start: ifst.start_state,
+        }
+    }
+}