+++ rustfst/src/fst_impls/mod.rs
@@ -1,3 +1,5 @@
-pub(crate) mod vector;
+pub(crate) mod const_fst;
+pub(crate) mod vector_fst;
 
-pub use self::vector::VectorFst;
+pub use self::const_fst::ConstFst;
+pub use self::vector_fst::VectorFst;