+++ rustfst/src/fst_impls/vector/mod.rs
@@ -1,6 +1,0 @@
-pub use self::vector_fst::VectorFst;
-
-mod mutable_fst;
-mod test;
-mod text_parser;
-pub(crate) mod vector_fst;