+++ rustfst/src/fst_impls/vector_fst/expanded_fst.rs
@@ -1,0 +1,9 @@
+use crate::fst_impls::VectorFst;
+use crate::fst_traits::ExpandedFst;
+use crate::semirings::Semiring;
+
+impl<W: 'static + Semiring> ExpandedFst for VectorFst<W> {
+    fn num_states(&self) -> usize {
+        self.states.len()
+    }
+}