+++ rustfst/src/test_data/text_fst/text_fst_001.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 use crate::arc::Arc;
-use crate::fst_impls::vector::vector_fst::VectorFstState;
+use crate::fst_impls::vector_fst::VectorFstState;
 use crate::fst_impls::VectorFst;
 use crate::parsers::text_fst::{FinalState, ParsedTextFst, Transition};
 use crate::semirings::{ProbabilityWeight, Semiring};